fix: resolve incorrect file cleanup in progress view due to path handling - move filterExistingFiles to workspaceFileUtils and fix relative path handling
